+++ fn/node-name.js
@@ -1,5 +1,3 @@
-'use strict';
-
 /**
  * _Extend jQuery methods_
  * @module
@@ -11,6 +9,7 @@
 // ----------------------------------------
 
 (function (window, $) {
+	'use strict';
 	/**
 	 * Returns `nodeName` property of DOM element  in lowercase.
 	 * Useful when you can get unknown elements.