+++ CHANGELOG.md
@@ -11,6 +11,13 @@
 
 ---
 
+## 1.2.0 [2017-09-20]
+
+- Add index bundle 
+- Change `'use strict'` level
+
+---
+
 ## 1.1.1 [2017-09-20]
 
 - Correction in [`$.fn.hasInitedKey()`](./README.md#fnhasinitedkey-key--setkey) method